requirements.txt
djangorestframework
django
pyyaml
uritemplate
mysqlclient
python-dotenv
django-environ
django-db-multitenant